// libs/components/src/molecules/inputs/radio/enum.ts
export enum RADIO_BUTTON_SIZE {
    SM = 'sm',
    MD = 'md',
    LG = 'lg',
  }
  
  export enum RADIO_BUTTON_VARIANT {
    PRIMARY = 'primary',
    ERROR = 'error',
    WARNING = 'warning',
  }
  
  export enum LABEL_SIZE {
    SM = 'sm',
    MD = 'md',
    LG = 'lg',
  }
  
  export enum STATUS_MESSAGE {
    SUCCESS = 'success',
    ERROR = 'error',
    WARNING = 'warning',
  }
  